+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.15.0
 
 - Add the ability to connect points visually
+- Add support for multi-selections by holding down <kbd>CMD</kbd> during click- or lasso-selections
 - Fix an issue with detecting points when using variable point sizes
 
 **Breaking changes:**